Use an off_t to store the difference between two off_t's.  Patch from Miquel van Smoorenburg.

+++ storage/cnfs/cnfs.c
@@ -1042,7 +1042,7 @@ TOKEN cnfs_store(const ARTHANDLE article, const STORAGECLASS class) {
     static int		iovcnt;
     int			tonextblock;
     CNFSEXPIRERULES	*metaexprule;
-    size_t		left;
+    off_t		left;
 
     for (metaexprule = metaexprulestab; metaexprule != (CNFSEXPIRERULES *)NULL; metaexprule = metaexprule->next) {
 	if (metaexprule->class == class)
